feat(subscriptions): Refactor premium plan creation to utilize subscription_service; update Paystack plan creation method

+++ backend/apps/subscriptions/services/paystack_service.py
@@ -30,7 +30,7 @@ class PaystackService:
         self,
         name: str,
         interval: str,
-        amount: Decimal,
+        amount: int,
         currency: Optional[str] = None,
         description: Optional[str] = None,
     ) -> Dict[str, Any]:
